Move `DowntimeListItem` to `Icinga\Module\Icingadb\Widget\ItemList`
